# [WIP] Pipelines - Reduce write lock contention