… src/System.IO.Pipelines/src/System/IO/Pipelines/PipeOperationState.cs