@stephentoub 2019-02-12
Doesn't this mean that a segment which might otherwise have been available for the continuation is now much less likely to be?  Is that not a problem at scale?