@benaadams 2019-02-12
> Is that not a problem at scale?

Pool size is 16; segments are 4096 bytes so that would affect many small writes without Flush larger than 60kB (15 segments * 4096b) or larger if the ArrayPool is being used.

If you are doing single shot large writes then it would be 16 writes without a Flush (say 1MB x 16 = 16MB) that would hit this.

On the flip size holding the lock which is for both reader and writer and delaying the signaling of the continuation is more of an issue as `ResetMemory` is pretty chunky and needs to be called for every block returned; as well as then either returing the data via ArrayPool.Return (or ConcurrentQueue for Kestrel); so the writer wouldn't be able to make progress during this time (currently)

e.g. (called for every block returned)
```asm
; Assembly listing for method BufferSegment:ResetMemory():this
; Lcl frame size = 40

G_M401_IG01:
       57                   push     rdi
       56                   push     rsi
       4883EC28             sub      rsp, 40
       488BF1               mov      rsi, rcx

G_M404_IG02:
       488B5628             mov      rdx, gword ptr [rsi+40]
       48B9B8D075C9FA7F0000 mov      rcx, 0x7FFAC975D0B8
       E8546A615F           call     CORINFO_HELP_ISINSTANCEOFINTERFACE
       4885C0               test     rax, rax
       7417                 je       SHORT G_M404_IG03
       488BC8               mov      rcx, rax
       49BB081536C9FA7F0000 mov      r11, 0x7FFAC9361508
       3909                 cmp      dword ptr [rcx], ecx
       FF1552B89AFF         call     [IDisposable:Dispose():this]
       EB49                 jmp      SHORT G_M404_IG04

G_M404_IG03:
       488B5628             mov      rdx, gword ptr [rsi+40]
       48B9020D50C9FA7F0000 mov      rcx, 0x7FFAC9500D02
       E845274F5F           call     CORINFO_HELP_ISINSTANCEOFARRAY
       488BF8               mov      rdi, rax
       4885FF               test     rdi, rdi
       742E                 je       SHORT G_M404_IG04
       48B9003F45C9FA7F0000 mov      rcx, 0x7FFAC9453F00
       BA77000000           mov      edx, 119
       E8B96C5F5F           call     CORINFO_HELP_CLASSINIT_SHARED_DYNAMICCLASS
       48B9E02E951095010000 mov      rcx, 0x19510952EE0
       488B09               mov      rcx, gword ptr [rcx]
       488BD7               mov      rdx, rdi
       4533C0               xor      r8d, r8d
       3909                 cmp      dword ptr [rcx], ecx
       E89795FFFF           call     TlsOverPerCoreLockedStacksArrayPool`1:Return(ref,bool):this

G_M405_IG04:
       33C0                 xor      rax, rax
       48894608             mov      gword ptr [rsi+8], rax
       33C0                 xor      rax, rax
       48894610             mov      qword ptr [rsi+16], rax
       33C0                 xor      rax, rax
       33D2                 xor      edx, edx
       488D4E18             lea      rcx, bword ptr [rsi+24]
       488901               mov      gword ptr [rcx], rax
       895108               mov      dword ptr [rcx+8], edx
       89510C               mov      dword ptr [rcx+12], edx
       33C0                 xor      rax, rax
       48894628             mov      gword ptr [rsi+40], rax
       48894630             mov      gword ptr [rsi+48], rax
       33C0                 xor      eax, eax
       894638               mov      dword ptr [rsi+56], eax
       33C0                 xor      rax, rax
       33D2                 xor      edx, edx
       4883C640             add      rsi, 64
       488906               mov      gword ptr [rsi], rax
       895608               mov      dword ptr [rsi+8], edx
       89560C               mov      dword ptr [rsi+12], edx

G_M407_IG05:
       4883C428             add      rsp, 40
       5E                   pop      rsi
       5F                   pop      rdi
       C3                   ret      

; Total bytes of code 197, prolog size 6 for method BufferSegment:ResetMemory():this
```