@pakrym 2019-02-12
Is there a way to make segment pool non-blocking? It doesn't have to be strict, loosing some segments to GC is acceptable.