@benaadams 2019-02-12
> Is there a way to make segment pool non-locking?

Easiest way is to use the [ConcurrentQueueSegment](https://github.com/dotnet/corefx/blob/ee3d69a9bacf585194b35de30045b64fcd3d7d77/src/Common/src/CoreLib/System/Collections/Concurrent/ConcurrentQueueSegment.cs) though its a bit chunky. /cc @stephentoub