@benaadams 2019-02-12
Though its single reader single writer, so might be able to do something simplier