@benaadams 2019-02-12
SingleProducerConsumerQueue?