@benaadams 2019-02-13
Made it lock-free and interleaved the Resets with the Returns