@benaadams 2019-02-13
Updated summary

> This change changes the segment pool to be lock-free so the whole pipe doesn't need to be locked to return them; interleaves ResetMemory and return; eagerly making the BufferSegments available and triggers completion prior to returning the segments.