@benaadams 2019-02-13
@stephentoub  does the conversion of `SingleProducerSingleConsumerQueue` to `SingleProducerSingleConsumerPool` look correct?