@benaadams 2019-02-13
Don't normally trust wrk on loopback from WSL to Windows but it does look significant

Before
```
root@ben-laptop:~# wrk -c 256 -t 4 -s ./wrk/scripts/pipeline.lua http://127.0.0.1:8080 -- 16
Pipeline depth: 16
Pipeline depth: 16
Pipeline depth: 16
Pipeline depth: 16
Running 10s test @ http://127.0.0.1:8080
  4 threads and 256 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.48ms    1.57ms  27.21ms   71.52%
    Req/Sec   149.17k    25.56k  213.80k    65.15%
  5954176 requests in 10.06s, 516.73MB read
Requests/sec: 591708.30
Transfer/sec:     51.35MB
```
After
```
root@ben-laptop:~# wrk -c 256 -t 4 -s ./wrk/scripts/pipeline.lua http://127.0.0.1:8080 -- 16
Pipeline depth: 16
Pipeline depth: 16
Pipeline depth: 16
Pipeline depth: 16
Running 10s test @ http://127.0.0.1:8080
  4 threads and 256 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.24ms    1.41ms  27.67ms   70.76%
    Req/Sec   164.85k    26.73k  259.68k    67.25%
  6580848 requests in 10.10s, 571.11MB read
Requests/sec: 651848.76
Transfer/sec:     56.57MB
```