@benaadams 2019-02-22
Removed the lock-free segment pool as `SingleProducerSingleConsumerQueue` is kinda chunky and it can be a future optimization.

Reduced the lock in the write path further following discussion in https://github.com/dotnet/corefx/pull/35484