@benaadams 2019-02-23
Fixed potential race between setting writing active and `_writingHead` being set to null (e.g. need to check it again after setting state to writing as it may have been set to null in the mean time)