@benaadams 2019-02-23
*note race condition was in this PR, not in current version